Nettoyage : utilisation de nombres relatifs (signés) pour les tailles des tableaux.
Ainsi nous pouvons mieux affirmer si oui ou non une taille est valide
(>= 0). Avec des size_t il faudrait vérifier -1, -2, -3, etc.

+++ bibliotheques/objets/import_objet.cc
@@ -113,7 +113,7 @@ static void lis_polygone(AdaptriceCreationObjet *adaptrice, std::istringstream &
 				index_polygones.data(),
 				index_coords_uv.data(),
 				index_normaux.data(),
-				index_polygones.size());
+				static_cast<long>(index_polygones.size()));
 }
 
 static void lis_objet(AdaptriceCreationObjet *adaptrice, std::istringstream &is)
